Changed the library to a class that needs to be inherited instead of AppCompatActivity
