+++ myLibrary/src/main/java/com/katyrin/mylibrary/AbsMyLibraryActivity.kt
@@ -1,0 +1,53 @@
+package com.katyrin.mylibrary
+
+import android.content.Intent
+import android.net.Uri
+import android.widget.Toast
+import androidx.appcompat.app.AppCompatActivity
+import io.reactivex.rxjava3.disposables.CompositeDisposable
+
+abstract class AbsMyLibraryActivity : AppCompatActivity() {
+
+    private val disposable: CompositeDisposable = CompositeDisposable()
+    private val activityLauncher = registerForActivityResult(MainActivityContract()) { intent ->
+        if (intent != null) takePersistableUriPermission(intent)
+    }
+
+    protected fun createNewActivity(requestText: String, stateListener: (String) -> Unit) {
+        subscribeState(stateListener)
+        startActivity(
+            LibraryActivity.getIntent(this, requestText)
+        )
+    }
+
+    private fun subscribeState(stateListener: (String) -> Unit) {
+        disposable.clear()
+        disposable.add(
+            RxBus.subscribe { stateListener(it) }
+        )
+    }
+
+    protected fun callDialogSelectionImages() {
+        activityLauncher.launch(null)
+    }
+
+    private fun takePersistableUriPermission(data: Intent) {
+        val uri = data.data ?: return
+        val takeFlags = data.flags.and(Intent.FLAG_GRANT_READ_URI_PERMISSION)
+            .or(Intent.FLAG_GRANT_WRITE_URI_PERMISSION)
+
+        try {
+            contentResolver.takePersistableUriPermission(uri, takeFlags)
+            setImageByUri(uri)
+        } catch (exception: SecurityException) {
+            Toast.makeText(this, exception.localizedMessage, Toast.LENGTH_SHORT).show()
+        }
+    }
+
+    abstract fun setImageByUri(uri: Uri)
+
+    override fun onDestroy() {
+        disposable.clear()
+        super.onDestroy()
+    }
+}
