--- myLibrary/src/main/java/com/katyrin/mylibrary/MyLibrary.kt
@@ -1,6 +1,0 @@
-package com.katyrin.mylibrary
-
-interface MyLibrary {
-    fun createNewActivity(requestText: String, stateListener: (String) -> Unit)
-    fun callDialogSelectionImages()
-}
